pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "content-strategy-agent"
version = "0.1.0"
description = "LangGraph skeleton for Instagram content strategy multi-agent workflow"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langgraph>=0.2.39",
  "langchain-core>=0.3.0",
  "apify-client>=1.8.1",
  "python-dotenv>=1.0.1",
  "google-genai>=1.13.0",
  "google-search-results>=2.4.2",
  "cerebras-cloud-sdk>=1.0.0",
  "pydantic>=2.0.0",
  "streamlit>=1.39.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "orchestrator*", "tools*"]
